+++ test/minitest_helper.rb
@@ -7,7 +7,6 @@ SimpleCov.start 'test_frameworks' do
   enable_coverage :branch
 end
 
-require 'yajl'
 require 'minitest'
 require 'minitest/autorun'